Pushing 2017-2018 research work
The aim was to define/shift points from a given matrix into the Brillouin zone defined by the lattice vectors given in the matrix's bxsf file.

Edge-casing for some metals was a work-in-progress, specifically when the Fermi Surface touched the surface of/left holes along the Brillouin zone.

A number of approachs to solve this edge-casing were attempted (e.g. lines 496-505, giving padding around Brillouin zone to cap holes), but nothing worked reliably with Marching Cubes to produce perfectly closed surfaces that could be 3D printed.